atualização

+++ src/main/java/kleberlz/libraryapi/service/AutorService.java
@@ -6,19 +6,24 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import kleberlz.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import kleberlz.libraryapi.model.Autor;
 import kleberlz.libraryapi.repository.AutorRepository;
+import kleberlz.libraryapi.repository.LivroRepository;
+import kleberlz.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 	
 	private final AutorRepository repository;
+	private final AutorValidator validator;
+	private final LivroRepository livroRepository;
 	
-	public AutorService(AutorRepository repository) {
-		this.repository = repository;
-	}
 	
 	public Autor salvar(Autor autor) {
+		validator.validar(autor);
 		return repository.save(autor);
 	}
 	
@@ -26,6 +31,7 @@ public class AutorService {
 		if(autor.getId() == null) {
 			throw new IllegalArgumentException("Para atualizar é necessário que o autor já esteja salvo na base.");
 		}
+		validator.validar(autor);
 		repository.save(autor);
 	}
 	
@@ -34,6 +40,10 @@ public class AutorService {
 	}
 
 	public void deletar(Autor autor) {
+		if(possuiLivro(autor)) {
+			throw new OperacaoNaoPermitidaException(
+					"Não é permitido excluir um Autor que possuí livros cadastrados!");
+		}
 		repository.delete(autor);
 	}
 	
@@ -51,5 +61,10 @@ public class AutorService {
 		}
 		
 		return repository.findAll();
+	}
+	
+	public boolean possuiLivro(Autor autor){
+		return livroRepository.existsByAutor(autor);
+		
 	}
 }
